requirements.txt

numpy
pandas

scikit-learn
joblib

transformers
sentence-transformers
torch

librosa
soundfile

opencv-python
Pillow

tensorflow
keras

spotipy
python-dotenv

streamlit
SpeechRecognition
pydub

soundfile


tqdm
